change(go): aggregate go codes into one file
