+++ go/app/main.go
@@ -210,6 +210,29 @@ func (i *itemRepositoryDB) Insert(ctx context.Context, item *Item) error {
 	return nil
 }
 
+// middleware functions
+func simpleCORSMiddleware(next http.Handler, origin string, methods []string) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		w.Header().Set("Access-Control-Allow-Origin", origin)
+		w.Header().Set("Access-Control-Allow-Methods", strings.Join(methods, ","))
+		w.Header().Set("Access-Control-Allow-Headers", "*")
+
+		if r.Method == "OPTIONS" {
+			w.WriteHeader(http.StatusOK)
+			return
+		}
+
+		next.ServeHTTP(w, r)
+	})
+}
+
+func simpleLoggerMiddleware(next http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		slog.Info("request received", "method", r.Method, "path", r.URL.Path, "remote_addr", r.RemoteAddr, "user_agent", r.UserAgent())
+		next.ServeHTTP(w, r)
+	})
+}
+
 func main() {
 	// set up logger
 	logger := slog.New(slog.NewJSONHandler(os.Stderr, nil))
